pyproject.toml
[build-system]
requires = ["setuptools>=61.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mongopy"
version = "0.0.2"
description = "Simple Package to manage operations with MongoDB Database"
readme = "README.md"
authors = [{ name = "Sagar Motghare", email = "sagarmotghare@proton.me" }]
license = { file = "LICENSE" }
classifiers = [
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
]
keywords = ["mongopy", "pymongo", "mongo"]
dependencies = ["pymongo", "pandas", "certifi"]
requires-python = ">=3.7"

[project.optional-dependencies]
build = ["build", "pymongo", "certifi", "pandas"]
dev = ["pymongo", "certifi", "pandas", "pytest", "python-dotenv"]

# [project.scripts]
# realpython = "reader.__main__:main"

[project.urls]
repository = "https://github.com/sagarmotghare/mongopy"
# documentation = "https://realpython.com/pypi-publish-python-package/"


[tool.bumpver]
current_version = "0.0.1"
version_pattern = "MAJOR.MINOR.PATCH"
commit_message = "bump version {old_version} -> {new_version}"
commit = true
tag = true
push = false

[tool.bumpver.file_patterns]
"pyproject.toml" = ['current_version = "{version}"', 'version = "{version}"']
"src/mongopy/__init__.py" = ["{version}"]

[tool.isort]
# profile = "black"
# import_heading_stdlib = "Standard library imports"
# import_heading_thirdparty = "Third party imports"
# import_heading_firstparty = "Reader imports"

[tool.mypy]
strict = true

[[tool.mypy.overrides]]
# module = "feedparser"
# ignore_missing_imports = true